Add PerfCounters support for Aarch64
For the moment only define NeoverseN1, which is what I'm testing on.
The scheme probably works on other microachitectures also, as long
as they support LSE, but we should test those one by one.
